// src/app/interfaces/experiences.ts
export interface Experiences {
  en:{
    title: string
    company: string
    from: string
    to: string
    details: {
      description: string
      location: string
      responsibilities: string[]
      link: string
    }
  },
  ba: {
    title: string
    company: string
    from: string
    to: string
    details: {
      description: string
      location: string
      responsibilities: string[]
      link: string
    }
  }
}